add delete collection option

+++ src/lib/trpc/routes/collections.ts
@@ -1,6 +1,7 @@
 import { createTRPCRouter, protectedProcedure } from '$lib/trpc/t';
 import { prisma } from '$lib/server/prisma';
 import { z } from 'zod';
+import { TRPCError } from '@trpc/server';
 
 export const collections = createTRPCRouter({
 	getUserCollections: protectedProcedure.query(({ ctx: { userId } }) =>
@@ -26,7 +27,13 @@ export const collections = createTRPCRouter({
 				})
 		),
 
-	deleteCollection: protectedProcedure.input(z.string()).mutation(async ({ input: id }) => {
-		await prisma.collection.delete({ where: { id } });
-	})
+	deleteCollection: protectedProcedure
+		.input(z.string())
+		.mutation(async ({ input: id, ctx: { userId } }) => {
+			const collection = await prisma.collection.findUniqueOrThrow({ where: { id } });
+
+			if (collection.ownerId !== userId) throw new TRPCError({ code: 'UNAUTHORIZED' });
+
+			await prisma.collection.delete({ where: { id } });
+		})
 });
